MonsterChase #34 - Player and Monster Dead State Changes: - src/animations.h: added empty distructor for player animations - src/monster.cpp: previor_state set to dead at construction time to force enter() execution; removed, leaked, default animation - src/monstersm.cpp/h: animation is now a common property of all states that is destryed and overridden at costruction were needed by a state; refactoring - src/player.cpp/h: enter() and exit() state execution support added; current_state is introduced and set to dead by default - src/playerqtviews.h: minor refactoring - src/playersm.cc/h: introducing _animation property in the state base class but initialize it only were needed (state normal, on_damage, dead) - tests/tests.pro: including into test project new objects

+++ src/playerqtviews.h
@@ -40,7 +40,7 @@ protected:
 
 private:
     QColor color[2];
-    int color_idx=0;
+    unsigned short color_idx=0;
     PlayerModel* model;
 
 };
